Add Execution Statistics to fog .

+++ NodeManager/Abstraction/IAppSettings.cs
@@ -8,6 +8,6 @@ namespace FogEnvironment.NodeManager.Abstraction
     {
         FogEnvironmentConfigurationModel FogEnvironmentModel { get; }
         TasksVolume TasksVolume { get; }
-        static double SientificNotationPower { get; set; }
+        static short SientificNotationPower { get; set; }
     }
 }
